Shift to nivo charts for intro page
