+++ frontend/src/data/nivo/common.ts
@@ -1,0 +1,8 @@
+export const serverToNivoData = (
+  name: string,
+  values: number[],
+  index: string[]
+) => ({
+  id: name,
+  data: values.map((y, i) => ({ x: index[i], y })),
+});
