+++ frontend/src/data/nivo/index.ts
@@ -1,0 +1,2 @@
+export * from './daily-load';
+export * from './daily-marginal';
